Initial shape of the game
There are many ways to represent the game of life.  We chose to model the
board as a sequence of sequences of cells, where the outer sequence represents
the rows, and the inner sequence represents the cells in each column.

We decided to make the board finite - i.e. it has edges and corners, rather than
wrapping round.  This makes the logic a little simpler (which is good for the
purposes of explaining the property-testing approach we took).

We used a sealed trait to represent a cell's state (Alive or Dead).

Finally, we decided to make the game itself a case class with a `next` method
which will give the next 'tick' of the game.