Working on the first test
The first test case for the game is fairly simple - a dead cell stays
dead if it has fewer than 3 living neighbours.  However, we know we
wnt to property test this, so we begin with a fairly simple premise.
Let's test the middle cell of a 3x3 board.

We generate a board of arbitrary cells, specifically set the middle
cell to be Dead, then check that calling `next` on the game leaves
the cell Dead.

This is the starting point of the test, but it's not complete yet.
We haven't actually made sure the middle cell has fewer than 3
neighbours!  We'll address that next.