More work on our first test
We want to make sure our test is correct - we say we're looking for
cases where the middle cell has fewer than 3 living neighbours, so
we put a `whenever` clause in our test to ensure this.  However, if
that's all we did, we'd end up throwing away a lot of test cases.

To combat this, we introduce a `frequency`.  Using this, we can tell
the `frequencyCell` generator how often we want Dead or Alive cells.
For this test, we want mainly Dead cells, so we supply it a ratio of
2:1.  We can tune this later if needs be.

Now our test setup is asserting the right thing, but it's still not
_really_ generic enough.  It's only testing cells that have all their
neighbours; it isn't testing edge or corner cells.  We could cover these
in separate tests, but actually their **behaviour** is exactly the same -
they should stay dead if they have fewer than 3 living neighbours.

What we really want to do is generate an arbitrary board, pick a cell,
and check that its behaviour is as expected.  Let's do that next.