Introduced `set` method
What happened there?!

When we looked at making our test more arbitrary, we realised a couple
of things.  Making the size arbitrary was easy, but setting an
arbitrary cell's value was a bit clunky.  Our test itself had a fair
bit of logic to it, and that was a sign that the interface of our
GameOfLife wasn't really helping us.

So, we introduced a `set` method on GameOfLife.

This might seem like writing production code just for the purposes
of our tests.  In fact, it's driving the design of our code through
our tests.  It's clear that the `set` method will be useful for our
tests, but it seemed highly likely that it'd be very useful for the
real code later.

Next, we'll refactor our test to use this new method and be more arbitrary.