Making our test arbitrary
Now that we have our `set` method, we can make the test arbitrary.
To do this we generate an arbitrary size, create a board of that
size with arbitrary cells, and pick arbitrary x and y
coordinates which gets set to being Dead.  We can then test
that this cell stays Dead on the next tick.

However, now that our board is bigger, finding how many living
neighbours our cell has is a bit trickier. The approach we take is
to look for the 3x3 grid centered on our target cell; but of course
some of these won't exist if our cell is on an edge or a corner.  So
we lit them into Options and discard any `None`'s.

This bit of code - finding a cell's neighbours - immediately looks
like another really useful piece of code, and one that our production
code is likely to use.  Likewise, the code to get a cell's value when
we don't know if its coordinates are on the board or not looks like
a candidate to split out.