New test - cells stay Dead with more than 3 living neighbours
I chose to put this test here to show a point - usually we'd move
on to something that drives new behaviour, but we wanted to look at
an interesting point here.  When we
add this test we realise that actually, these two behaviours are
simple subsets of a wider behaviour - that Dead cells stay Dead
if they don't have exactly 3 living neighbours.

Sometimes it's useful to split behaviours up into separate cases
like this.  Sometimes it's clearer to combine them.  Think about
what explains the intended behaviour more clearly.  In this case I
think we'll combine them, which we'll do next.