Dead cells come alive with 3 living neighbours
This is the first test we've added that means we change our main
implementation.  Again, our test is exhaustive - arbitrary board,
arbitrary coordinates and the baheviour always holds, whether it's
a cell on the edge, a corner, or in the middle.  We simply tune the
frequency of our test data to ensure we get lots of valid cases
(i.e. lots of cells with 3 living neighbours).

Note also that in our two tests we've covered the entire problem
space of what to do with dead cells.  That's pretty cool!

The implementation simply maps over each cell and checks if it's
Dead and has 3 living neighbours.  If that's true, it makes sure
the cell becomes Alive.  Otherwise it leaves it alone.

Time to add another behaviour.