Our final test for living cells
We add the last test - that living cells die if they have more than
3 living neighbours - and refactor the code a little.  There's more
refactoring that could be done of course (e.g. the repeated
`.count(_ == Alive)` looks like one candidate), but we stop here as
the point was in the tests.

Look back at the rules of the Game of Life.  There are 4 explicit
rules, and there's one _implicit_ rule as well - that Dead cells
stay Dead unless they have 3 neighbours.  Look back at our tests -
we have one for each of the 4 explicit rules and one for the
implicit one.

This is really powerful.  It means our tests - our Spec file -
*actually** specifies the rules of the game.  Someone reading our
tests can understand the entire problem clearly.  The fact we use
a property testing engine and generated data also means we get
lots of confidence that our code works in all sorts of cases.