Refector
All of the tests for the game of life itself use very similar generators,
so we extract that to a common place.

We also introduce a simple filter to count the number of living neighbours,
since we use this filter in numerous places.

Finally, we tune the frequency of our generators slighly.